requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
llama-cpp-python==0.2.23
pydantic==2.4.2 